refactor: product and refund double type change to bigdecimal type

+++ src/main/java/com/paymentteamproject/domain/refund/dto/RefundCreateResponse.java
@@ -3,6 +3,7 @@ package com.paymentteamproject.domain.refund.dto;
 import com.paymentteamproject.domain.refund.consts.RefundStatus;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -10,12 +11,12 @@ public class RefundCreateResponse {
 
     private final Long refundId;
     private final String paymentId;
-    private final double amount;
+    private final BigDecimal amount;
     private final RefundStatus status;
     private final String reason;
     private final LocalDateTime refundAt;
 
-    public RefundCreateResponse(Long refundId, String paymentId, double amount, RefundStatus status, String reason, LocalDateTime refundAt) {
+    public RefundCreateResponse(Long refundId, String paymentId, BigDecimal amount, RefundStatus status, String reason, LocalDateTime refundAt) {
         this.refundId = refundId;
         this.paymentId = paymentId;
         this.amount = amount;
